+++ knowde/complex/system/domain/navi/errors.py
@@ -1,5 +1,1 @@
 """errors."""
-
-
-class HeadingNotFoundError(Exception):
-    """見出しが見つからない."""
