+++ knowde/complex/system/domain/sysnet/errors.py
@@ -3,3 +3,7 @@
 
 class UnResolvedTermError(Exception):
     """用語解決がまだ."""
+
+
+class HeadingNotFoundError(Exception):
+    """見出しが見つからない."""
